+++ NannysLM/src/app/components/hero/hero.css
@@ -63,9 +63,9 @@
 
 .cta-buttons {
   display: flex;
-  flex-direction: column;
-  gap: 1rem;
-  max-width: 400px;
+  gap: 1.5rem;
+  max-width: 600px;
+  flex-wrap: wrap;
 }
 
 .btn-search {
@@ -85,6 +85,7 @@
   box-shadow: 0 8px 25px rgba(108, 99, 255, 0.3);
   position: relative;
   overflow: hidden;
+  white-space: nowrap;
 }
 
 .btn-search::before {
@@ -120,11 +121,19 @@
   box-shadow: 0 8px 25px rgba(255, 105, 180, 0.4);
   position: relative;
   overflow: hidden;
+  white-space: nowrap;
 }
 
 .btn-learn:hover {
   transform: translateY(-3px);
   box-shadow: 0 12px 35px rgba(255, 105, 180, 0.5);
+}
+
+/* Desktop: make "Conocenos" wider than the primary CTA */
+@media (min-width: 1025px) {
+  .btn-learn {
+    min-width: 250px;
+  }
 }
 
 /* Hero Image */
@@ -564,8 +573,18 @@
   }
 
   .cta-buttons {
+    flex-direction: column;
     align-items: center;
+    justify-content: center;
     margin: 0 auto;
+    gap: 1rem;
+    width: 100%;
+    max-width: 400px;
+  }
+
+  .btn-search,
+  .btn-learn {
+    width: 100%;
   }
 
   .features {
