+++ NannysLM/src/app/components/hero/hero.html
@@ -7,7 +7,7 @@
     <p class="hero-description">Conectamos familias con cuidadoras verificadas y confiables. Tu tranquilidad es nuestra prioridad.</p>
     
     <div class="cta-buttons">
-      <button class="btn-search">
+      <button class="btn-search" routerLink="/register">
         <svg width="20" height="20" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
           <circle cx="11" cy="11" r="8" stroke="currentColor" stroke-width="2"/>
           <path d="m21 21-4.35-4.35" stroke="currentColor" stroke-width="2"/>
