+++ NannysLM/src/app/components/hero/hero.ts
@@ -1,8 +1,10 @@
 import { Component } from '@angular/core';
+import { RouterModule } from '@angular/router';
 
 @Component({
   selector: 'app-hero',
   standalone: true,
+  imports: [RouterModule],
   templateUrl: './hero.html',
   styleUrls: ['./hero.css']
 })
