+++ NannysLM/src/app/components/login/login.html
@@ -15,6 +15,12 @@
   <!-- Sección derecha con el formulario de login -->
   <div class="form-section">
     <div class="form-container">
+      <!-- Botón de regreso para móvil -->
+      <button class="mobile-back-button" routerLink="/">
+        <svg width="24" height="24" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
+          <path d="M15 18L9 12L15 6" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"/>
+        </svg>
+      </button>
 
       <h2>Bienvenido de vuelta</h2>
       <p class="subtitle">Inicia sesión para encontrar la niñera perfecta</p>
