+++ NannysLM/src/app/components/register/register.html
@@ -15,7 +15,13 @@
   <!-- Sección derecha con el formulario de registro -->
   <div class="form-section">
     <div class="form-container">
-            
+      <!-- Botón de regreso para móvil -->
+      <button class="mobile-back-button" routerLink="/">
+        <svg width="24" height="24" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
+          <path d="M15 18L9 12L15 6" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"/>
+        </svg>
+      </button>
+      
       <h2>Crear cuenta</h2>
       <p class="subtitle">Únete a nuestra comunidad de familias</p>
