+++ NannysLM/src/app/pages/nanny/nanny-dashboard.component.css
@@ -39,7 +39,7 @@ app-notifications-panel ::ng-deep {
 .nanny-layout {
   display: flex;
   flex-direction: column;
-  min-height: 100vh;
+  height: 100vh;
   font-family: 'Segoe UI', Tahoma, Geneva, Verdana, sans-serif;
   background-color: var(--gray-50);
 }
@@ -129,6 +129,15 @@ app-notifications-panel ::ng-deep {
   display: flex;
   flex: 1;
   overflow: hidden;
+  height: 100%;
+}
+
+/* Sidebar fijo */
+app-sidebar {
+  display: flex;
+  flex-shrink: 0;
+  height: 100%;
+  overflow: hidden;
 }
 
 /* Contenido principal */
@@ -136,9 +145,12 @@ app-notifications-panel ::ng-deep {
   flex: 1;
   padding: 2rem;
   overflow-y: auto;
+  overflow-x: hidden;
   background: #f8fafc;
   width: 100%;
   box-sizing: border-box;
+  display: flex;
+  flex-direction: column;
 }
 
 .content-header {
@@ -162,7 +174,7 @@ app-notifications-panel ::ng-deep {
 
 /* Tarjeta de bienvenida hero */
 .welcome-hero-card {
-  background: linear-gradient(135deg, #e31b7e 0%, #ff6b9d 100%);
+  background: #e31b7e;
   border-radius: 20px;
   padding: 2.5rem;
   margin-bottom: 2.5rem;
@@ -179,12 +191,6 @@ app-notifications-panel ::ng-deep {
   width: 200%;
   height: 200%;
   background: radial-gradient(circle, rgba(255, 255, 255, 0.1) 0%, transparent 70%);
-  animation: float 6s ease-in-out infinite;
-}
-
-@keyframes float {
-  0%, 100% { transform: translateY(0px) rotate(0deg); }
-  50% { transform: translateY(-20px) rotate(5deg); }
 }
 
 .hero-content {
