+++ NannysLM/src/app/pages/nanny/nanny-dashboard.component.html
@@ -179,68 +179,6 @@
           </div>
         </div>
 
-        <!-- Sección de acceso rápido -->
-        <div class="quick-access-section">
-          <div class="section-header">
-            <h3>Acceso Rápido</h3>
-            <p>Funciones más utilizadas</p>
-          </div>
-          
-          <div class="quick-access-grid">
-            <button class="quick-access-item" (click)="currentView = 'services'">
-              <div class="access-icon primary">
-                <svg width="24" height="24" viewBox="0 0 24 24" fill="currentColor" xmlns="http://www.w3.org/2000/svg">
-                  <path d="M19 3H5C3.9 3 3 3.9 3 5V19C3 20.1 3.9 21 5 21H19C20.1 21 21 20.1 21 19V5C21 3.9 20.1 3 19 3ZM19 19H5V8H19V19Z"/>
-                </svg>
-              </div>
-              <div class="access-content">
-                <h4>Mis Servicios</h4>
-                <p>Ver programación completa</p>
-              </div>
-              <div class="access-arrow">
-                <svg width="16" height="16" viewBox="0 0 24 24" fill="currentColor" xmlns="http://www.w3.org/2000/svg">
-                  <polyline points="9,18 15,12 9,6" stroke="currentColor" stroke-width="2" fill="none"/>
-                </svg>
-              </div>
-            </button>
-
-            <button class="quick-access-item" (click)="setServiceFilter('upcoming')">
-              <div class="access-icon secondary">
-                <svg width="24" height="24" viewBox="0 0 24 24" fill="currentColor" xmlns="http://www.w3.org/2000/svg">
-                  <path d="M12 2C13.65 2 15 3.35 15 5C15 6.65 13.65 8 12 8C10.35 8 9 6.65 9 5C9 3.35 10.35 2 12 2Z"/>
-                </svg>
-              </div>
-              <div class="access-content">
-                <h4>Próximos Servicios</h4>
-                <p>Trabajos programados</p>
-              </div>
-              <div class="access-arrow">
-                <svg width="16" height="16" viewBox="0 0 24 24" fill="currentColor" xmlns="http://www.w3.org/2000/svg">
-                  <polyline points="9,18 15,12 9,6" stroke="currentColor" stroke-width="2" fill="none"/>
-                </svg>
-              </div>
-            </button>
-
-            <button class="quick-access-item" (click)="setServiceFilter('past')">
-              <div class="access-icon tertiary">
-                <svg width="24" height="24" viewBox="0 0 24 24" fill="currentColor" xmlns="http://www.w3.org/2000/svg">
-                  <path d="M22 11.08V12a10 10 0 1 1-5.93-9.14"/>
-                  <polyline points="22,4 12,14.01 9,11.01" stroke="currentColor" stroke-width="2" fill="none"/>
-                </svg>
-              </div>
-              <div class="access-content">
-                <h4>Historial</h4>
-                <p>Servicios completados</p>
-              </div>
-              <div class="access-arrow">
-                <svg width="16" height="16" viewBox="0 0 24 24" fill="currentColor" xmlns="http://www.w3.org/2000/svg">
-                  <polyline points="9,18 15,12 9,6" stroke="currentColor" stroke-width="2" fill="none"/>
-                </svg>
-              </div>
-            </button>
-          </div>
-        </div>
-
         <!-- Últimos servicios completados (si los hay) -->
         <div class="recent-activity-section" *ngIf="services.past.length > 0">
           <div class="section-header">
@@ -400,7 +338,7 @@
       <!-- Vista Notificaciones -->
       <div *ngIf="currentView === 'notificaciones'" class="notifications-view">
         <div class="content-header">
-          <h2>📬 Notificaciones</h2>
+          <h2>Mis Notificaciones</h2>
           <p>Gestiona tus notificaciones y actualizaciones</p>
         </div>
