+++ NannysLM/src/app/pages/nanny/service-details.component.css
@@ -190,41 +190,20 @@
   box-shadow: 0 4px 12px rgba(0, 0, 0, 0.05);
 }
 
-.back-button {
-  display: inline-flex;
-  align-items: center;
-  gap: 0.5rem;
-  padding: 0.5rem 1rem;
-  background: transparent;
-  color: #64748b;
-  border: 1px solid #e2e8f0;
-  border-radius: 8px;
-  font-size: 0.9rem;
-  font-weight: 500;
-  cursor: pointer;
-  transition: all 0.3s ease;
-  margin-bottom: 1.5rem;
-}
-
-.back-button:hover {
-  background: #f8fafc;
-  border-color: #cbd5e1;
-  color: #475569;
-}
-
 .service-title-section {
   display: flex;
   align-items: center;
   justify-content: space-between;
   flex-wrap: wrap;
-  gap: 1rem;
+  gap: 1.5rem;
 }
 
 .service-title-section h1 {
-  font-size: 2rem;
+  font-size: 1.8rem;
   color: #1f2937;
   margin: 0;
   flex: 1;
+  word-break: break-word;
 }
 
 .service-badge {
@@ -232,6 +211,7 @@
   border-radius: 8px;
   font-size: 0.9rem;
   font-weight: 600;
+  white-space: nowrap;
 }
 
 .service-badge.pending {
@@ -297,7 +277,7 @@
 /* Details Grid */
 .details-grid {
   display: grid;
-  grid-template-columns: repeat(auto-fit, minmax(300px, 1fr));
+  grid-template-columns: repeat(auto-fit, minmax(280px, 1fr));
   gap: 1.5rem;
   margin-bottom: 2rem;
 }
@@ -326,18 +306,20 @@
 
 .card-header svg {
   flex-shrink: 0;
+  color: #E31B7E;
 }
 
 .card-header h3 {
   margin: 0;
-  font-size: 1.1rem;
+  font-size: 1rem;
   color: #1f2937;
+  font-weight: 600;
 }
 
 .card-content {
   display: flex;
   flex-direction: column;
-  gap: 1rem;
+  gap: 1.25rem;
 }
 
 .detail-row {
@@ -349,8 +331,9 @@
 
 .detail-row .label {
   color: #64748b;
-  font-size: 0.9rem;
+  font-size: 0.85rem;
   font-weight: 500;
+  flex-shrink: 0;
 }
 
 .detail-row .value {
@@ -358,11 +341,12 @@
   font-size: 0.95rem;
   font-weight: 600;
   text-align: right;
+  word-break: break-word;
 }
 
 .detail-row .value.amount {
   color: #E31B7E;
-  font-size: 1.2rem;
+  font-size: 1.1rem;
 }
 
 .address-text {
@@ -409,6 +393,7 @@
   position: sticky;
   bottom: 2rem;
   z-index: 10;
+  margin-bottom: 2rem;
 }
 
 .action-btn {
@@ -416,14 +401,14 @@
   align-items: center;
   justify-content: center;
   gap: 0.5rem;
-  padding: 1rem 2rem;
+  padding: 0.75rem 1.5rem;
   border: none;
   border-radius: 12px;
-  font-size: 1rem;
+  font-size: 0.95rem;
   font-weight: 600;
   cursor: pointer;
   transition: all 0.3s ease;
-  min-width: 180px;
+  min-width: 160px;
 }
 
 .action-btn:disabled {
@@ -484,30 +469,210 @@
 /* Responsive */
 @media (max-width: 1024px) {
   .main-content {
-    margin-left: 0;
+    padding: 100px 1.5rem 2rem 1.5rem;
   }
 
   .details-grid {
-    grid-template-columns: 1fr;
+    grid-template-columns: repeat(2, 1fr);
+  }
+
+  .top-bar {
+    padding: 0 1.5rem;
   }
 }
 
 @media (max-width: 768px) {
   .main-content {
-    padding: 1rem;
+    padding: 100px 1rem 2rem 1rem;
+  }
+
+  .service-header {
+    padding: 1.5rem;
+  }
+
+  .service-title-section {
+    flex-direction: column;
+    align-items: flex-start;
+    gap: 1rem;
   }
 
   .service-title-section h1 {
     font-size: 1.5rem;
   }
 
+  .service-badge {
+    align-self: flex-start;
+  }
+
+  .details-grid {
+    grid-template-columns: 1fr;
+  }
+
+  .urgency-banner,
+  .info-banner,
+  .error-banner {
+    flex-direction: column;
+  }
+
   .actions-section {
     flex-direction: column;
-    bottom: 1rem;
+    gap: 1rem;
+    position: static;
+    padding: 1.5rem 1rem;
+    margin: 2rem -1rem -2rem -1rem;
   }
 
   .action-btn {
     width: 100%;
+    min-width: 100%;
+  }
+
+  .page-title {
+    font-size: 1.1rem;
+  }
+
+  .top-bar {
+    padding: 0 1rem;
+  }
+
+  .detail-card {
+    padding: 1rem;
+  }
+
+  .instructions-section {
+    padding: 1.5rem;
+  }
+}
+
+@media (max-width: 480px) {
+  .main-content {
+    padding: 90px 0.75rem 1rem 0.75rem;
+  }
+
+  .service-header {
+    padding: 1rem;
+  }
+
+  .service-title-section h1 {
+    font-size: 1.2rem;
+  }
+
+  .service-badge {
+    padding: 0.4rem 0.75rem;
+    font-size: 0.8rem;
+  }
+
+  .detail-card {
+    padding: 0.75rem;
+  }
+
+  .card-header {
+    margin-bottom: 0.75rem;
+    gap: 0.5rem;
+  }
+
+  .card-header h3 {
+    font-size: 1rem;
+  }
+
+  .detail-row {
+    flex-direction: column;
+    align-items: flex-start;
+    gap: 0.25rem;
+  }
+
+  .detail-row .label {
+    font-size: 0.8rem;
+  }
+
+  .detail-row .value {
+    text-align: left;
+    font-size: 0.9rem;
+  }
+
+  .actions-section {
+    padding: 1rem 0.75rem;
+    margin: 1.5rem -0.75rem -2rem -0.75rem;
+    border-radius: 0;
+  }
+
+  .action-btn {
+    padding: 0.75rem 1rem;
+    font-size: 0.9rem;
+    min-width: auto;
+  }
+
+  .top-bar {
+    padding: 0 0.75rem;
+    height: 65px;
+  }
+
+  .back-button {
+    padding: 0.4rem 0.75rem;
+    font-size: 0.8rem;
+  }
+
+  .page-title {
+    font-size: 0.95rem;
+  }
+
+  .urgency-banner,
+  .info-banner,
+  .error-banner {
+    padding: 0.75rem;
+    gap: 0.5rem;
+  }
+
+  .urgency-banner svg,
+  .info-banner svg,
+  .error-banner svg {
+    width: 20px;
+    height: 20px;
+  }
+
+  .urgency-banner strong,
+  .info-banner strong,
+  .error-banner strong {
+    font-size: 0.9rem;
+  }
+
+  .urgency-banner p,
+  .info-banner p,
+  .error-banner p {
+    font-size: 0.8rem;
+  }
+
+  .instructions-section {
+    padding: 1rem;
+  }
+
+  .instructions-section h3 {
+    font-size: 1rem;
+    margin-bottom: 0.75rem;
+  }
+
+  .instructions-section p {
+    font-size: 0.9rem;
+  }
+
+  .address-text {
+    font-size: 0.9rem;
+  }
+
+  .modal-content {
+    width: 95%;
+  }
+
+  .modal-header {
+    padding: 1rem;
+  }
+
+  .modal-body {
+    padding: 1rem;
+  }
+
+  .modal-footer {
+    padding: 1rem;
   }
 }
