+++ NannysLM/src/app/pages/nanny/service-details.component.html
@@ -58,13 +58,6 @@
       <div *ngIf="service && !isLoading && !acceptSuccess && !completeSuccess" class="service-details-container">
         <!-- Header del servicio -->
         <div class="service-header">
-          <button class="back-button" (click)="declineService()">
-            <svg width="20" height="20" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
-              <path d="M19 12H5M5 12L12 19M5 12L12 5" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"/>
-            </svg>
-            Volver
-          </button>
-          
           <div class="service-title-section">
             <h1>{{ service.title }}</h1>
             <span class="service-badge" [class.pending]="service.status === 'pending'" [class.assigned]="service.nanny_id">
