+++ NannysLM/src/app/shared/components/notifications-panel/notifications-panel.component.html
@@ -1,7 +1,7 @@
 <!-- Header -->
 <div *ngIf="showHeader" style="margin-bottom: 1.5rem;">
   <h3 style="margin: 0 0 0.5rem 0; font-size: 1.25rem; color: #1f2937; font-weight: 600;">
-    {{ title }}
+    Notificaciones
     <span style="font-size: 0.875rem; color: #6b7280; margin-left: 0.5rem;">
       ({{ getTotalCount() }})
     </span>
@@ -9,10 +9,11 @@
 </div>
 
 <!-- Filtros -->
-<div *ngIf="showFilters" style="display: flex; gap: 0.75rem; margin-bottom: 1rem;">
+<div *ngIf="showFilters" style="display: flex; gap: 0.75rem; margin-bottom: 1rem;" role="group" aria-label="Filtros de notificaciones">
   <button 
     type="button"
     (click)="notificationsView = 'unread'"
+    [attr.aria-pressed]="notificationsView === 'unread'"
     [style.background]="notificationsView === 'unread' ? '#e31b7e' : '#f3f4f6'"
     [style.color]="notificationsView === 'unread' ? 'white' : '#374151'"
     style="padding: 0.5rem 1rem; border: none; border-radius: 6px; cursor: pointer; font-weight: 500; transition: all 0.3s ease; font-size: 0.9375rem;">
@@ -21,6 +22,7 @@
   <button 
     type="button"
     (click)="notificationsView = 'all'"
+    [attr.aria-pressed]="notificationsView === 'all'"
     [style.background]="notificationsView === 'all' ? '#e31b7e' : '#f3f4f6'"
     [style.color]="notificationsView === 'all' ? 'white' : '#374151'"
     style="padding: 0.5rem 1rem; border: none; border-radius: 6px; cursor: pointer; font-weight: 500; transition: all 0.3s ease; font-size: 0.9375rem;">
@@ -35,7 +37,7 @@
 </div>
 
 <!-- Lista de notificaciones -->
-<div *ngIf="!isLoadingNotifications" style="display: flex; flex-direction: column; gap: 0.75rem;">
+<div *ngIf="!isLoadingNotifications" style="display: flex; flex-direction: column; gap: 0.75rem;" role="feed" aria-label="Lista de notificaciones">
   <!-- Sin notificaciones -->
   <div *ngIf="getFilteredNotifications().length === 0" 
     style="padding: 2rem; text-align: center; background: #f9fafb; border-radius: 8px; color: #6b7280;">
@@ -45,8 +47,10 @@
   </div>
 
   <!-- Notificaciones -->
-  <div #notificationCard
+  <article #notificationCard
     *ngFor="let notification of getFilteredNotifications()" 
+    [attr.role]="'article'"
+    [attr.aria-label]="notification.title"
     [style.opacity]="notification.is_read ? '0.7' : '1'"
     style="padding: 1rem; border: 1px solid #e5e7eb; border-radius: 8px; background: white; cursor: pointer; transition: all 0.3s ease; border-left: 4px solid;"
     [style.border-left-color]="getNotificationColor(notification.type)"
@@ -59,6 +63,11 @@
         <!-- Título -->
         <h4 style="margin: 0 0 0.25rem 0; font-size: 1rem; color: #1f2937; font-weight: 600;">
           {{ notification.title }}
+          <span *ngIf="!notification.is_read" 
+            class="sr-only" 
+            aria-label="no leída">
+            (sin leer)
+          </span>
         </h4>
         
         <!-- Mensaje -->
@@ -68,33 +77,46 @@
         
         <!-- Fecha y estado -->
         <div style="display: flex; justify-content: space-between; align-items: center; font-size: 0.8125rem; color: #9ca3af;">
-          <span>{{ formatDate(notification.created_at) }}</span>
+          <time [dateTime]="notification.created_at">{{ formatDate(notification.created_at) }}</time>
           <span *ngIf="!notification.is_read" 
             style="display: inline-block; width: 8px; height: 8px; background: #e31b7e; border-radius: 50%; margin-left: 0.5rem;" 
-            title="No leída"></span>
+            title="No leída"
+            role="status"
+            aria-label="notificación no leída"></span>
         </div>
       </div>
 
       <!-- Botones de acción -->
-      <div style="display: flex; gap: 0.5rem; flex-shrink: 0;">
+      <div style="display: flex; gap: 0.5rem; flex-shrink: 0;" role="toolbar" aria-label="Acciones">
         <button 
           *ngIf="!notification.is_read"
           type="button"
           (click)="markNotificationAsRead(notification, $event)"
           title="Marcar como leída"
-          style="padding: 0.375rem 0.75rem; background: #f0f9ff; color: #0369a1; border: 1px solid #bae6fd; border-radius: 6px; cursor: pointer; font-size: 0.8125rem; font-weight: 500; transition: all 0.2s ease;">
-          ✓ Leer
+          aria-label="Marcar como leída"
+          style="padding: 0.375rem 0.75rem; background: #f0f9ff; color: #0369a1; border: 1px solid #bae6fd; border-radius: 6px; cursor: pointer; font-size: 0.8125rem; font-weight: 500; transition: all 0.2s ease; display: flex; align-items: center; gap: 0.35rem;">
+          <svg width="14" height="14" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg" aria-hidden="true">
+            <polyline points="20 6 9 17 4 12" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"/>
+          </svg>
+          Leer
         </button>
         <button 
           type="button"
           (click)="deleteNotification(notification.id, $event)"
           title="Eliminar"
-          style="padding: 0.375rem 0.75rem; background: #fee2e2; color: #dc2626; border: 1px solid #fecaca; border-radius: 6px; cursor: pointer; font-size: 0.8125rem; font-weight: 500; transition: all 0.2s ease;">
-          🗑️
+          aria-label="Eliminar notificación"
+          style="padding: 0.375rem 0.75rem; background: #fee2e2; color: #dc2626; border: 1px solid #fecaca; border-radius: 6px; cursor: pointer; font-size: 0.8125rem; font-weight: 500; transition: all 0.2s ease; display: flex; align-items: center; gap: 0.35rem;">
+          <svg width="14" height="14" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg" aria-hidden="true">
+            <polyline points="3 6 5 6 21 6" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"/>
+            <path d="M19 6v14a2 2 0 0 1-2 2H7a2 2 0 0 1-2-2V6m3 0V4a2 2 0 0 1 2-2h4a2 2 0 0 1 2 2v2" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"/>
+            <line x1="10" y1="11" x2="10" y2="17" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"/>
+            <line x1="14" y1="11" x2="14" y2="17" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"/>
+          </svg>
+          Eliminar
         </button>
       </div>
     </div>
-  </div>
+  </article>
 </div>
 
 <!-- Estilos globales -->
